+++ internal/testdata/testtwo/config.go
@@ -1,0 +1,5 @@
+package testtwo
+
+type Config struct {
+	Bb int
+}