+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### v2.8.4
+* moved event entities to event-lib
+* fix jsonschema cross package definition collision
 ### v2.8.3
 * bootstrap: fix empty connection string
 * bootstrap: always send module requirements